requirements.txt
streamlit~=1.11.0
toml~=0.10.2
google-auth~=2.6.2